requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic-settings
python-multipart
psycopg2-binary
